+++ src/main/java/com/wss/websoso/memo/MemoController.java
@@ -1,5 +1,8 @@
 package com.wss.websoso.memo;
 
+import com.wss.websoso.memo.dto.MemoDetailGetResponse;
+import com.wss.websoso.memo.dto.MemoUpdateRequest;
+import com.wss.websoso.memo.dto.MemosGetResponse;
 import java.security.Principal;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
