+++ src/main/java/com/wss/websoso/memo/dto/MemoCreateResponse.java
@@ -1,4 +1,4 @@
-package com.wss.websoso.memo;
+package com.wss.websoso.memo.dto;
 
 public record MemoCreateResponse(
         Boolean isAvatarUnlocked
