+++ src/main/java/com/wss/websoso/user/UserService.java
@@ -6,8 +6,10 @@ import com.wss.websoso.avatarLine.AvatarLine;
 import com.wss.websoso.avatarLine.AvatarLineRepository;
 import com.wss.websoso.config.jwt.JwtProvider;
 import com.wss.websoso.config.jwt.UserAuthentication;
+import com.wss.websoso.user.dto.UserInfoGetResponse;
+import com.wss.websoso.user.dto.UserLoginRequest;
 import com.wss.websoso.userAvatar.UserAvatarRepository;
-import com.wss.websoso.userNovel.UserAvatarsGetResponse;
+import com.wss.websoso.userNovel.dto.UserAvatarsGetResponse;
 import com.wss.websoso.userNovel.UserNovelRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
