+++ src/main/java/com/wss/websoso/user/dto/UserInfoGetResponse.java
@@ -1,8 +1,9 @@
-package com.wss.websoso.user;
+package com.wss.websoso.user.dto;
 
 import com.wss.websoso.avatar.Avatar;
 import com.wss.websoso.avatarLine.AvatarLine;
-import com.wss.websoso.userNovel.UserAvatarsGetResponse;
+import com.wss.websoso.user.User;
+import com.wss.websoso.userNovel.dto.UserAvatarsGetResponse;
 import java.util.List;
 
 public record UserInfoGetResponse(
