+++ src/main/java/com/wss/websoso/user/dto/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package com.wss.websoso.user;
+package com.wss.websoso.user.dto;
 
 public record UserLoginRequest(
         String Authorization
