+++ src/main/java/com/wss/websoso/userNovel/UserNovelService.java
@@ -11,6 +11,12 @@ import com.wss.websoso.platform.Platform;
 import com.wss.websoso.platform.PlatformRepository;
 import com.wss.websoso.user.User;
 import com.wss.websoso.user.UserRepository;
+import com.wss.websoso.userNovel.dto.SosoPickGetResponse;
+import com.wss.websoso.userNovel.dto.SosoPicksGetResponse;
+import com.wss.websoso.userNovel.dto.UserNovelCreateRequest;
+import com.wss.websoso.userNovel.dto.UserNovelMemoAndInfoGetResponse;
+import com.wss.websoso.userNovel.dto.UserNovelUpdateRequest;
+import com.wss.websoso.userNovel.dto.UserNovelsResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
